+++ frontend/src/components/common/Text/MediaQueryText.ts
